+++ LogicBuilder.EntityFrameworkCore.SqlServer/Repositories/IContextRepository.cs
@@ -19,7 +19,7 @@ namespace LogicBuilder.EntityFrameworkCore.SqlServer.Repositories
             where TModel : IBaseModel
             where TData : class, IBaseData;
 
-        Task<TModelReturn> QueryAsync<TModel, TData, TModelReturn, TDataReturn>(Expression<Func<IQueryable<TModel>, TModelReturn>> queryFunc, SelectExpandDefinition? selectExpandDefinition = null)
+        Task<TModelReturn> QueryAsync<TModel, TData, TModelReturn, TDataReturn>(Expression<Func<IQueryable<TModel>, TModelReturn>> queryFunc, SelectExpandDefinition? selectExpandDefinition = null)//NOSONAR -  in this case, reducing the number of generic parameters adds more complexity and would not be beneficial to readability.
             where TModel : IBaseModel
             where TData : class, IBaseData;